+++ src/Conversation/Conversation.inte.js
@@ -5,8 +5,8 @@ import { shallow, mount } from 'enzyme';
 import { expect } from 'chai';
 
 describe('<Conversation />', () => {
-  it('renders one p element', () => {
-    const wrapper = shallow(<Conversation />);
+  it('should render ', () => {
+    const wrapper = mount(<Conversation />);
     expect(wrapper.find('p')).to.have.length(1);
     expect(wrapper.find('p').text()).to.equal('Hello World!');
   });