+++ src/Conversation/Conversation.js
@@ -1,18 +1,8 @@
 import React from 'react';
 
-// function getStyles(props, context) {
-//   const styles = {
-//   };
-//
-//   return styles;
-// }
-
-const Conversation = (props) => {
-  return (
+const Conversation = () =>
     <div>
-      <p>Hello World and {props.name}!</p>
+      <p>Hello World!</p>
     </div>
-  );
-};
 
 export default Conversation;